% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xreg.R
\name{build_xreg_backtest}
\alias{build_xreg_backtest}
\title{Build backtest xreg using only training data to forecast exog}
\usage{
build_xreg_backtest(y_ts, ex_ts, L, train_end)
}
\arguments{
\item{y_ts}{response ts}

\item{ex_ts}{exogenous ts}

\item{L}{lag}

\item{train_end}{c(year, month) end point of training}
}
\value{
list(x_tr, x_te, h_bt, Ttr)
}
\description{
Build backtest xreg using only training data to forecast exog
}
