% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xreg.R
\name{build_xreg_main}
\alias{build_xreg_main}
\title{Build main xreg for forward forecasting (y_t ~ x_{t-L})}
\usage{
build_xreg_main(y_ts, ex_ts, L, h_out)
}
\arguments{
\item{y_ts}{response ts}

\item{ex_ts}{exogenous ts (observed history)}

\item{L}{lag to align x_{t-L}}

\item{h_out}{forecast horizon}
}
\value{
list(x_tr, x_te) matrices
}
\description{
Build main xreg for forward forecasting (y_t ~ x_{t-L})
}
