% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_exog.R, R/xreg.R
\name{forecast_exog}
\alias{forecast_exog}
\title{Forecast exogenous series}
\usage{
forecast_exog(series, h)

forecast_exog(series, h)
}
\arguments{
\item{series}{numeric or ts}

\item{h}{integer forecast horizon}
}
\value{
ts mean forecast

ts mean forecast
}
\description{
Forecast exogenous series

Forecast an exogenous series with ARIMA
}
