% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{invlog}
\alias{invlog}
\title{Inverse of log with positive shift}
\usage{
invlog(z, shift)
}
\arguments{
\item{z}{numeric vector (log scale)}

\item{shift}{positive shift used in log(x + shift)}
}
\value{
numeric vector on original scale
}
\description{
Inverse of log with positive shift
}
