% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gif.R
\name{make_forecast_gif}
\alias{make_forecast_gif}
\title{Create animation (GIF) of Actual / Backtest Forecast / Future Forecast}
\usage{
make_forecast_gif(
  combined,
  pg_name,
  bt_df = NULL,
  backtest_start_date = as.Date("2023-01-01"),
  duration = 8,
  fps = 20,
  width = 900,
  height = 500
)
}
\arguments{
\item{combined}{data.frame: columns date\if{html}{\out{<Date>}}, value\if{html}{\out{<numeric>}}, type {"Actual","Forecast"}}

\item{pg_name}{character label for title/filename}

\item{bt_df}{optional data.frame: date\if{html}{\out{<Date>}}, Actual, Forecast (from backtest)}

\item{backtest_start_date}{Date for dashed vline}

\item{duration}{seconds}

\item{fps}{frames per second}

\item{width, height}{pixel size of GIF}
}
\value{
list(gif=path, preview=png path)
}
\description{
Create animation (GIF) of Actual / Backtest Forecast / Future Forecast
}
