% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{select_features}
\alias{select_features}
\title{Elastic-Net feature selection}
\usage{
select_features(dt, y, xs, alpha = 0.5, k = 10)
}
\arguments{
\item{dt}{data.table with response and candidate features}

\item{y}{character name of response column (e.g., "Va_log")}

\item{xs}{character vector of candidate feature names (e.g., CSI1:CSI85)}

\item{alpha}{mixing parameter in Elastic-Net (0=LASSO, 1=Ridge)}

\item{k}{number of CV folds}
}
\value{
character vector of selected feature names
}
\description{
Elastic-Net feature selection
}
